Updated code to control motor port #0 on PCA9685 Board

+++ pca9685.h
@@ -25,6 +25,6 @@ private:
     int address_;
     int file_descriptor_;
 
-    void write_reg(int reg, int value);
-    int read_reg(int reg);
+    void    write_reg(int reg, int value);
+    int8_t  read_reg(int reg);
 };
